At some moments you just need to kill something anything will do....
Errors like these are just that frustrating

(And don't take this as a death threat. If you do you'll be wasting my time and that of the police and are already not doing what they have to do, so I'd like to avoid that miserable excuse for a law and order uphold organisation).