Auto color adaption
It goes like this. All devices using colors use a 0-255 scale. No exceptions.
But some (...... CENSORED ......) thought it was a good idea to use a 0-1 scale (only to have it calculated back to 0-255 in the underlying API and forcing me to devide all color calls by 255 or my code will no longer work, causing nothing but a lot of slow down in the longer run).
All in all, I don't know whose idea it is, but my module is now able to detect this automatically and use the correct function, and am I in luck I was always too lazy to write love.graphics.setColor all the time..... Damn!